Revert "Manage HMAC instances per thread"
This reverts commit a0e6ac10ed780dc7882f9c79da916a632a20f56c.
